Refactor interceptors into interceptors package
(cherry picked from commit 1382c90ce11cf74373c9e03e08b8531e613afe19)
